ML/requirements.txt
grpcio==1.71.0
grpcio-tools==1.71.0
joblib==1.4.2
numpy==2.0.2
pandas==2.2.3
protobuf==5.29.4
python-dateutil==2.9.0.post0
pytz==2025.1
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.1
